[p10] refactor: break apart debug function to modularization
