+++ lib/dbug/dbug.h
@@ -1,0 +1,7 @@
+#ifndef DEBUG_H
+#define DEBUG_H
+
+// Função para debugar
+void debug_print(const char *format, ...);
+
+#endif
